+++ src/main/java/org/betonquest/betonquest/compatibility/mmogroup/mmolib/MythicLibIntegrator.java
@@ -4,17 +4,17 @@ import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.compatibility.Integrator;
 
 @SuppressWarnings("PMD.CommentRequired")
-public class MMOLibIntegrator implements Integrator {
+public class MythicLibIntegrator implements Integrator {
 
     private final BetonQuest plugin;
 
-    public MMOLibIntegrator() {
+    public MythicLibIntegrator() {
         plugin = BetonQuest.getInstance();
     }
 
     @Override
     public void hook() {
-        plugin.registerConditions("mmostat", MMOLibStatCondition.class);
+        plugin.registerConditions("mmostat", MythicLibStatCondition.class);
     }
 
     @Override