+++ src/main/java/org/betonquest/betonquest/events/CommandEvent.java
@@ -23,7 +23,8 @@ public class CommandEvent extends QuestEvent {
         super(instruction, true);
         staticness = true;
         persistent = true;
-        final String string = instruction.getInstruction().trim();
+        String string = instruction.getInstruction();
+        string = string.replace("conditions:", "").trim();
         // Split commands by | but allow one to use \| to represent a pipe character
         final String[] rawCommands = Arrays.stream(string.substring(string.indexOf(' ') + 1).split("(?<!\\\\)\\|"))
                 .map(s -> s.replace("\\|", "|"))