+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 - `npcrange` objective is triggered at wrong time
 - Citizens compatibility for not spawned NPCs
+- NotifyIOs are case-sensitive
 ### Security
 
 ## [1.12.1] - 2021-02-05