+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Citizens compatibility for not spawned NPCs
 - NotifyIOs are case-sensitive
 - all mmo objectives trigger for everyone
+- command event includes 'conditions:...' into the command
 ### Security
 
 ## [1.12.1] - 2021-02-05